Extract and Create User Interface class
