+++ src/main/java/Duke/Exception/DukeException.java
@@ -1,75 +1,44 @@
 package Duke.Exception;
 
+import Duke.UI.UserInterface;
+
 import java.io.IOException;
 
-public class DukeException extends Exception{
-    public static final String LINE = "____________________________________________________________";
 
+public class DukeException extends Exception{
     public static void emptyTodoException() {
-        System.out.println(LINE);
-        System.out.println("Your todo command does not have an argument!");
-        System.out.println("To input a valid todo command, type \"todo (argument)\"!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.emptyTodoMessage();
     }
 
     public static void invalidDeadlineException() {
-        System.out.println(LINE);
-        System.out.println("Your deadline command is invalid!");
-        System.out.println("To input a valid deadline command, type \"deadline (description) /by (deadline)\"!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.invalidDeadlineMessage();
     }
 
     public static void invalidEventException() {
-        System.out.println(LINE);
-        System.out.println("Your event command is invalid!");
-        System.out.println("To input a event command, type \"event (description) /at (when)\"!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.invalidEventMessage();
     }
 
     public static void invalidDoneException() {
-        System.out.println(LINE);
-        System.out.println("Your done command is invalid!");
-        System.out.println("To input a done command, type \"done (task index)\"!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.invalidDoneMessage();
     }
 
     public static void invalidDeleteException() {
-        System.out.println(LINE);
-        System.out.println("Your delete command is invalid!");
-        System.out.println("To input a delete command, type \"delete (task index)\"!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.invalidDeleteMessage();
     }
 
     public static void emptyTaskException() {
-        System.out.println(LINE);
-        System.out.println(" No Tasks here yet. Go include some tasks!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.emptyTaskMessage();
     }
 
     public static void invalidSaveFileException() {
-        System.out.println(LINE);
-        System.out.println("Your save file is invalid!");
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.invalidSaveFileMessage();
     }
 
     public static void createIOException(IOException ioException) {
-        System.out.println(LINE);
-        System.out.println("Something went wrong: " + ioException.getMessage());
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.createIOExceptionMessage(ioException);
     }
 
     public static void SaveIOException(IOException ioException) {
-        System.out.println(LINE);
-        System.out.println("Something went wrong: " + ioException.getMessage());
-        System.out.println(LINE);
-        System.out.print(System.lineSeparator());
+        UserInterface.SaveIOExceptionMessage(ioException);
     }
 }
